ai-service/requirements.txt
langchain
langgraph
pydantic
ipython
grandalf
python-dotenv
requests
langchain-groq
Flask